add move_base_swp actionlib interface
This patch adds the second actionlib interface that will be
used for sparse-waypoint plan. At present, the move_base_swp
action interface is the same as the original move_base interface
and it is the primary interface for talking to move_base.

The second interface is legacy and it's a simply proxy that
takes the move_base goal and relays it to the primary interface.
Future commits will change the type of the move_base_swp interface
and the legacy interface will do the adaptation, retaining
backwards compatibility.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>