use the MoveBaseSWP on main interface
Rework the main interface to accept the vector of poses
(waypoints) and adapt the legacy interface to forward
a single-element vector to the main interface when receiving
the goal.

This is only the interface adaptation. The planner does not
(yet) honor the waypoints. Instead it just plans to the
last waypoint. The actual rework of the planning loop
is yet to come, but this change sets the interface into
a stable state.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>