track progress through the global plan
whenever the robot moves, find the pose among plan waypoints
that is closest to the current robot pose and call it the
current robot position in the plan. This will be used to track
the waypoint the planner is pursuing and determine the start of the
sliding window for partial feeding into the local planner.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>