improve the search for the closest waypoint
If the plan makes a small step back (which typically
happens where sparse waypoints are stitched) the
search for the nearest way point may get stuck in the
bogus minimum (basically at the sparse waypoint).

To work around that problem enforce a minimum look-ahead
window when searching for nearest waypoints.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>