prune infeasable waypoints
If a robot is positioned such that a waypoint would
require it to go to it and come back to the same location
and the the total distance is within the short-term
planning window, the planner can end up
in an infinite loop of planning, immediately succeeding,
and re-planning the same path.

If such a situation happens, declare that the waypoint
is infeasable, prune it and try again.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>