eliminate deadlock condition in brake thread
Some functions may call applyBrakes, grabbing the
brake mutex with configuration mutex held. Grabbing
a configuration mutex in brake thread while holding
a the brake mutex will result in mutexes being grabbed
in the opposite order, which is a deadlock condition.

Eliminate the condition by releasing the brake lock
when resampling the parameters and then reacquire the lock.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>